#491: Generalize CustomWorldSerializationManager.
