+++ artemis-serialization/artemis-json-libgdx/src/test/java/com/artemis/managers/CustomJsonLibGdxWorldSerializationManagerTest.java
@@ -6,9 +6,8 @@ import com.artemis.annotations.Wire;
 import com.artemis.io.JsonArtemisSerializer;
 import com.artemis.io.SaveFileFormat;
 import com.artemis.utils.IntBag;
-import com.esotericsoftware.jsonbeans.Json;
-import com.esotericsoftware.jsonbeans.JsonSerializer;
-import com.esotericsoftware.jsonbeans.JsonValue;
+import com.badlogic.gdx.utils.Json;
+import com.badlogic.gdx.utils.JsonValue;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -20,17 +19,17 @@ import java.nio.charset.StandardCharsets;
 
 import static org.junit.Assert.*;
 
-public class CustomJsonWorldSerializationManagerTest extends AbstractCustomWorldSerializationManagerTest {
+public class CustomJsonLibGdxWorldSerializationManagerTest extends AbstractCustomWorldSerializationManagerTest {
 
 	@Override
 	protected WorldSerializationManager.ArtemisSerializer<?> createBackend(World world) {
 		JsonArtemisSerializer backend = new JsonArtemisSerializer(world);
 		backend.prettyPrint(true);
-		backend.register(AbstractCustomWorldSerializationManagerTest.SerializedSystem.class, new SerializedSystemSerializer(world));
+		backend.register(SerializedSystem.class, new SerializedSystemSerializer(world));
 		return backend;
 	}
 
-	public static class SerializedSystemSerializer implements JsonSerializer<SerializedSystem> {
+	public static class SerializedSystemSerializer implements Json.Serializer<SerializedSystem> {
 		private SerializedSystem system;
 
 		public SerializedSystemSerializer(World world) {
@@ -51,17 +50,4 @@ public class CustomJsonWorldSerializationManagerTest extends AbstractCustomWorld
 		}
 	}
 
-	public static class CustomSaveFormat extends SaveFileFormat {
-		public SerializedSystem serialized;
-		public DummySegment noSerializer;
-
-		public CustomSaveFormat(EntitySubscription es, String dummyString, int dummyNumber) {
-			super(es);
-			noSerializer = new DummySegment(dummyString, dummyNumber);
-		}
-
-		private CustomSaveFormat() {
-			super((IntBag)null);
-		}
-	}
 }
