+++ artemis-serialization/artemis-kryo/src/test/java/com/artemis/managers/CustomKryoWorldSerializationManagerTest.java
@@ -2,6 +2,7 @@ package com.artemis.managers;
 
 import com.artemis.*;
 import com.artemis.component.*;
+import com.artemis.io.JsonArtemisSerializer;
 import com.artemis.io.KryoArtemisSerializer;
 import com.artemis.io.SaveFileFormat;
 import com.artemis.utils.IntBag;
@@ -20,90 +21,15 @@ import java.nio.charset.StandardCharsets;
 
 import static org.junit.Assert.*;
 
-public class CustomKryoWorldSerializationManagerTest {
-	private WorldSerializationManager manger;
-	private AspectSubscriptionManager subscriptions;
-	private SerializedSystem serializedSystem;
-	private World world;
-	private EntitySubscription allEntities;
+public class CustomKryoWorldSerializationManagerTest extends AbstractCustomWorldSerializationManagerTest {
 
-	@Before
-	public void setup() {
-
-		world = new World(new WorldConfiguration()
-				.setSystem(SerializedSystem.class)
-				.setSystem(WorldSerializationManager.class));
-
-		world.inject(this);
+	@Override
+	protected WorldSerializationManager.ArtemisSerializer<?> createBackend(World world) {
 		KryoArtemisSerializer backend = new KryoArtemisSerializer(world);
 		backend.register(SerializedSystem.class, new SerializedSystemSerializer(world));
 		backend.register(CustomSaveFormat.class);
 		backend.register(CustomKryoWorldSerializationManagerTest.DummySegment.class);
-		manger.setSerializer(backend);
-
-		allEntities = subscriptions.get(Aspect.all());
-
-		world.process();
-		assertEquals(0, allEntities.getEntities().size());
-	}
-
-	@Test
-	public void custom_save_format_save_load() throws Exception {
-		serializedSystem.serializeMe = "dog";
-
-		byte[] save = save(allEntities, "a string", 420);
-		serializedSystem.serializeMe = "cat";
-
-		deleteAll();
-		assertEquals(0, allEntities.getEntities().size());
-
-		ByteArrayInputStream bais = new ByteArrayInputStream(save);
-		CustomSaveFormat load = manger.load(bais, CustomSaveFormat.class);
-
-		world.process();
-		assertEquals("DOG", serializedSystem.serializeMe);
-		assertEquals("DOG", load.serialized.serializeMe);
-		assertEquals("a string", load.noSerializer.text);
-		assertEquals(420, load.noSerializer.number);
-	}
-
-	private byte[] save(EntitySubscription subscription, String s, int i)
-			throws Exception {
-
-		SaveFileFormat save = new CustomSaveFormat(subscription, s, i);
-		ByteArrayOutputStream baos = new ByteArrayOutputStream(256);
-		manger.save(baos, save);
-		return baos.toByteArray();
-	}
-
-	private int deleteAll() {
-		IntBag entities = allEntities.getEntities();
-		int size = entities.size();
-		for (int i = 0; size > i; i++) {
-			world.delete(entities.get(i));
-		}
-
-		world.process();
-		return size;
-	}
-
-	public static class SerializedSystem extends BaseSystem {
-		public String serializeMe;
-
-		@Override
-		protected void processSystem() {}
-	}
-
-	public static class DummySegment {
-		public String text;
-		public int number;
-
-		public DummySegment(String text, int number) {
-			this.text = text;
-			this.number = number;
-		}
-
-		public DummySegment() {}
+		return backend;
 	}
 
 	public static class SerializedSystemSerializer extends Serializer<SerializedSystem> {
